*: remove some demo code and add LICENSE header
